+++ JoyboyCommunity/src/hooks/index.ts
@@ -1,20 +1,9 @@
+export * from './modals';
+export * from './nostr';
 export {useChainId} from './useChainId';
 export {useColor} from './useColor';
-export {useDialog} from './useDialog';
-export {useEditProfile} from './useEditProfile';
-export {useNote} from './useNote';
-export {useProfile} from './useProfile';
-export {useReactions} from './useReactions';
-export {useReplyNotes} from './useReplyNotes';
-export {useReposts} from './useReposts';
-export {useRootNotes} from './useRootNotes';
-export {useSendNote} from './useSendNote';
 export {useStyles} from './useStyles';
 export {useTheme} from './useTheme';
-export {useTipModal} from './useTipModal';
 export {useTips} from './useTips';
-export {useToast} from './useToast';
 export {useTransaction} from './useTransaction';
-export {useTransactionModal} from './useTransactionModal';
 export {useWaitConnection} from './useWaitConnection';
-export {useWalletModal} from './useWalletModal';
