+++ JoyboyCommunity/src/hooks/nostr/useEditProfile.ts
@@ -1,8 +1,8 @@
 import {NDKUserProfile} from '@nostr-dev-kit/ndk';
 import {useMutation} from '@tanstack/react-query';
 
-import {useNostrContext} from '../context/NostrContext';
-import {useAuth} from '../store/auth';
+import {useNostrContext} from '../../context/NostrContext';
+import {useAuth} from '../../store/auth';
 
 export const useEditProfile = () => {
   const {ndk} = useNostrContext();
