+++ JoyboyCommunity/src/hooks/nostr/useNote.ts
@@ -1,7 +1,7 @@
 import {NDKKind} from '@nostr-dev-kit/ndk';
 import {useQuery} from '@tanstack/react-query';
 
-import {useNostrContext} from '../context/NostrContext';
+import {useNostrContext} from '../../context/NostrContext';
 
 export type UseNoteOptions = {
   noteId: string;
