+++ JoyboyCommunity/src/hooks/nostr/useProfile.ts
@@ -1,6 +1,6 @@
 import {useQuery} from '@tanstack/react-query';
 
-import {useNostrContext} from '../context/NostrContext';
+import {useNostrContext} from '../../context/NostrContext';
 
 export type UseProfileOptions = {
   publicKey: string;
