+++ JoyboyCommunity/src/hooks/nostr/useReposts.ts
@@ -1,19 +1,19 @@
 import {NDKKind} from '@nostr-dev-kit/ndk';
 import {useInfiniteQuery} from '@tanstack/react-query';
 
-import {useNostrContext} from '../context/NostrContext';
+import {useNostrContext} from '../../context/NostrContext';
 
-export type UseReactionsOptions = {
+export type UseRepostsOptions = {
   authors?: string[];
   search?: string;
 };
 
-export const useReactions = (options?: UseReactionsOptions) => {
+export const useReposts = (options?: UseRepostsOptions) => {
   const {ndk} = useNostrContext();
 
   return useInfiniteQuery({
     initialPageParam: 0,
-    queryKey: ['reactions', options?.authors, options?.search],
+    queryKey: ['reposts', options?.authors, options?.search],
     getNextPageParam: (lastPage: any, allPages, lastPageParam) => {
       if (!lastPage?.length) return undefined;
 
@@ -23,15 +23,15 @@ export const useReactions = (options?: UseReactionsOptions) => {
       return pageParam;
     },
     queryFn: async ({pageParam}) => {
-      const notes = await ndk.fetchEvents({
-        kinds: [NDKKind.Reaction],
+      const reposts = await ndk.fetchEvents({
+        kinds: [NDKKind.Repost],
         authors: options?.authors,
         search: options?.search,
         until: pageParam || Math.round(Date.now() / 1000),
         limit: 20,
       });
 
-      return [...notes];
+      return [...reposts];
     },
     placeholderData: {pages: [], pageParams: []},
   });
