+++ JoyboyCommunity/src/hooks/nostr/useRootNotes.ts
@@ -1,7 +1,7 @@
 import {NDKKind} from '@nostr-dev-kit/ndk';
 import {useInfiniteQuery} from '@tanstack/react-query';
 
-import {useNostrContext} from '../context/NostrContext';
+import {useNostrContext} from '../../context/NostrContext';
 
 export type UseRootNotesOptions = {
   authors?: string[];
