+++ JoyboyCommunity/src/hooks/useTransaction.ts
@@ -1,7 +1,7 @@
 import {ContractWriteVariables, useContractWrite} from '@starknet-react/core';
 import {GetTransactionReceiptResponse} from 'starknet';
 
-import {useTransactionModal} from './useTransactionModal';
+import {useTransactionModal} from './modals/useTransactionModal';
 
 export const useTransaction = () => {
   const {show: showTransactionModal, hide: hideTransactionModal, shown} = useTransactionModal();
